Use "/builds/${PRODUCT}/${VERSION_NUMBER}/${BUILD_TYPE}" path for saving both master and stable builds.
This logic is now simpler and is possible since we decided to ditch following the "/builds/${PRODUCT}/nightlies/${VERSION_NUMBER}"
path convention which was anyways symlinked to "/builds/${PRODUCT}/${BUILD_TYPE}" and just l=make the later a proper folder with the right builds directly.

This way both build.ixsystems.com and download.freenas.org naming scheme is united.