Free FreeNAS from the shackles of 10/100 ethernet support.
Ticket: #16178

This commit removes the 10/100 ethernet drivers from the FreeNAS
kernel.  Since most of this hardware is over 10 years old it's a
pretty safe bet that it can't run FreeNAS anyways, being that
FreeNAS went 64 bit only over a year ago.

Note that I left rl support in because of the iconic comments
in if_rl.c, not because I consider it likely that anyone is running
FreeNAS with a 100mbps realtek card. (If you are, you should stop that :)

Note that this commit is not merely cleaning house for the sake of
cleaning house, as dtrace has a limit of 32K objects and FreeNAS is
just sneaking in under that.