split "git pull --rebase" into component parts
Sometimes "git pull --rebase" fails with:

    fatal: BUG: get_tempfile_fd() called for inactive object

(I cannot get it to fail for me.)  Recent changes to git made
pull a built-in C program rather than a shell script, and perhaps
the new reference locking code is being tickled badly.  So
for now, change this to

    git fetch && git rebase

which is essentially what the old shell script based "git pull"
would do here (though of course it had to do it the very long
way around, but we should be fine with this short way).