enable ACPI_DEBUG in DEBUG kernel
Several customers are experiencing mutex issues with ACPI on two
different motherboards so far.  Enable in-kernel ACPI_DEBUG for
the DEBUG kernel so that we can find out why and maybe work around
it.

(This is similar to d42d8919f7e9aee54f27dc4c030168b6edaf4d3a
except that it applies to multiple kernel configs)