Link fdescfs statically into the kernel.
We depend on fdescfs to make python behave.  I see no reason to
statically link other FS like tmpfs or nullfs, but keep fdescfs out.

There are some reports about odd kernel panics that happen on fdescfs
mount, but not on nullfs or tmpfs mounts, while the code is identical.
My only guess now is that it is somehow related to dynamic linkage.

Ticket:	#27315